% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_selection_shiny.R
\name{cut_selection_shiny}
\alias{cut_selection_shiny}
\title{A partir d'un repertoire contenant l'ensemble des png, selectionne les coupes et les copies dans path_output}
\usage{
cut_selection_shiny(
  path_raw,
  path_selected,
  path_processed,
  cut_list = c(106),
  process = FALSE
)
}
\arguments{
\item{path_raw}{Repertoire où sont stockées l'ensemble des coupes}

\item{path_selected}{Repertoire où vont etre stocké les coupes selectionnées}

\item{path_processed}{Repertoire où vont etre stocké les coupes selectionnées et pré-proccessées}

\item{cut_list}{Liste des slices à selectionner}

\item{process}{Bool si nous devons preprocesser ou pas les coupes selectionnées}
}
\value{
Retourne un datatable contenant les infos sur les coupes selectionnees
}
\description{
A partir d'un repertoire contenant l'ensemble des png, selectionne les coupes et les copies dans path_output
}
