% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_augment.R
\name{data_augment}
\alias{data_augment}
\title{Data augmentation d'un repertoire donnee}
\usage{
data_augment(
  nb_target,
  path_source,
  path_dest,
  data_ref,
  cat2augment,
  translat = 1.1,
  max_zoom = 1.1,
  max_angle = 5
)
}
\arguments{
\item{nb_target}{nombre d'image cible}

\item{path_source}{Chemin source des images}

\item{path_dest}{Chemin destination images augmentees}

\item{data_ref}{table de reference des images.}

\item{cat2augment}{Categorie a augmenter}

\item{translat}{Coef de translation max}

\item{max_zoom}{Coef de zoom max}

\item{max_angle}{Coef de rotation max}
}
\value{
data table
}
\description{
Data augmentation d'un repertoire donnee
}
