% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_image.R
\name{extend_image}
\alias{extend_image}
\title{Extrait les features d'une image unique}
\usage{
extend_image(img, width_max = 254, height_max = 254)
}
\arguments{
\item{img}{Chemin absolu de l'image}

\item{width_max}{Largeur cible}

\item{height_max}{Hauteur cible}
}
\value{
retourne un vecteur
}
\description{
Extrait les features d'une image unique
}
