% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_feature.R
\name{extract_feature}
\alias{extract_feature}
\title{Extrait les features d'une repertoire contenant des images}
\usage{
extract_feature(path_dir, data_ref, width = 256, height = 256, labelsExist = T)
}
\arguments{
\item{path_dir}{Repertoire contenant les images}

\item{data_ref}{Table de reference du dataset}

\item{width}{Largeur de l'image cible}

\item{height}{Hauteur de l'image cible}

\item{labelsExist}{Booléen si le label existe ou non}
}
\value{
Retourne une liste (X = feature_matrix, y = y)) si lalelsExist = T ou uniquement feature_matrix si lalelsExist = F
}
\description{
Extrait les features d'une repertoire contenant des images
}
