% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_list.R
\name{feature_list}
\alias{feature_list}
\title{Extrait les features d'une image unique}
\usage{
feature_list(path_img, width = 254, height = 254)
}
\arguments{
\item{path_img}{Chemin absolu de l'image}

\item{width}{Largeur cible}

\item{height}{Hauteur cible}
}
\value{
retourne un vecteur
}
\description{
Extrait les features d'une image unique
}
