% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRdata.R
\name{loadRData}
\alias{loadRData}
\title{Load Rdata in a R variable}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{char Rdata filename}
}
\value{
variable
}
\description{
Load Rdata in a R variable
}
